BasicParser

Description

Documentation

type ParserFunction a = String -> Maybe (a, String)

A ParserFunction function take a string a parameter, maybe returns a successfull value of type "a" tuppled with a string (the rest of the string)

data Parser a

A Parser of type "a" holds a ParserFunction of same type.

Constructors

Parser

Fields

runParser :: ParserFunction a

(<&>) :: Parser a -> Parser b -> Parser (a, b)

Calls "parser1" and calls "parser2" on what "parser1" didn't parse and returns a tupple of "parser1" and "parser2"

(<^>) :: Parser a -> Parser a -> Parser a

XOR parser

(<%>) :: Parser String -> Parser a -> Parser a

Calls "p1", and passing "p2" the parsed value. Returns the result of "p2"

parseHead :: Parser Char

Parse the first element of the stream

parseChar :: Char -> Parser Char

If the argument is the first element in the string, returns it Otherwise, returns Nothing

parseAnyChar :: String -> Parser Char

If one char of the first argument is the first element in the string, returns it Otherwise, returns Nothing

parseAndWith :: (a -> b -> c) -> Parser a -> Parser b -> Parser c

Calls <&> and applies "func" on the result

parseMany :: Parser a -> Parser [a]

While "p" doesn't returns Nothing, it is called on the rest Returns an array of parsed values. Never returns Nothing

parseSome :: Parser a -> Parser [a]

While "p" doesn't returns Nothing, it is called on the rest Returns an array of parsed values. Returns Nothing if the array is empty

parseUInt :: Parser Integer

Parse unsigned integer from string

parseInt :: Parser Integer

Parse signed integer from string

parseFloat :: Parser Float

Parse floating point number from string, returns Nothing if is NOT a floating point number

parseScope :: (Char, Char) -> Parser String

Extract what's in the parenthesis

parseWhitespaces :: Parser String

Parse anything that is a whitespace (using isSpace)

parseWhile :: (Char -> Bool) -> Parser String

Parse String While function returns true, Nothing if Nothing was parsed

parseIf :: (Char -> Bool) -> Parser Char

Parse String if function returns true

parseN :: Int -> Parser a -> Parser [a]

Run parser n times

parseString :: String -> Parser String

Parse the string from the stream if it exists

parseWord :: Parser String

Parse anything that is not a whitespace